feat: example consuming  useNotifiSubscriptionContext/NotifiClientContext (sui)
