+++ packages/notifi-react-example/src/NotifiContextWrapper/index.ts
@@ -2,3 +2,4 @@ export * from './SolanaNotifiContextWrapper';
 export * from './WalletConnectNotifiContextWrapper';
 export * from './AcalaNotifiContextWrapper';
 export * from './KeplrNotifiContextWrapper';
+export * from './SuiNotifiContextWrapper';
